84. Largest Rectangle in Histogram - O(N)
Idea: keep the stack of "previous i higher than h"; when h[currentI]
< stacktip.h, we pop the tip as {i, h}, such that all [i, currentI - 1]
are higher or equal to h, and area would be `(currentI - i) * h`. But
the i to push back is "previous i that is higher or equal to h".

Every i, will be pushed into, and popped out of the stack once, so the
time and space are both O(N).

A more concise answer: "left border" is actually "next stack tip",
so we can simply push currentI into stack. But complexity is the same.